requirements.txt
pandas==1.3.3
matplotlib==3.4.3
seaborn==0.11.2
mysql-connector-python==8.0.23
scikit-learn==0.24.2
statsmodels==0.12.2